backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
dspy-ai
langchain
langchain-groq
langchain-qdrant
langchain-huggingface
langchain-mcp-adapters
langgraph
qdrant-client
sympy
pyyaml
sentence-transformers
